+++ skbio/io/format/embed.py
@@ -225,18 +225,24 @@ def _objects_to_embed(objs, fh, include_embedding_pointer=True):
             if "dim" not in h5grp.attrs:
                 h5grp.attrs["dim"] = emb.shape[1]
 
-            # resize if necessary
-            if i > 0:
+            # resizing if necessary
+            if i > 0: # we don't need to resize for the first object
                 if include_embedding_pointer:
-                    if (len(arr) + idptr_fh[i - 1]) > max_idsize or (
-                        emb.shape[0] + embptr_fh[i - 1]
-                    ) > max_embsize:
-                        max_idsize = ceil(len(arr) + idptr_fh[i - 1]) * resize_by
-                        max_embsize = ceil(emb.shape[0] + embptr_fh[i - 1]) * resize_by
-                        resize = True
-                else:
-                    if len(arr) + idptr_fh[i - 1] > max_idsize:
-                        max_idsize = ceil(len(arr) + idptr_fh[i - 1] * resize_by)
+                    # Checking 4 conditions for possible resize:
+                    # if (new ID/embs + old ID/embs) > max_size
+                    if (len(arr) + idptr_fh[i - 1]) >= max_idsize or (
+                        emb.shape[0] + embptr_fh[i - 1]) >= max_embsize or (
+                        i >= len(idptr_fh) or (
+                        # check if adding obj exceeds id array size
+                        i >= embptr_fh.shape[0])):
+                        # check if adding obj exceeds embedding array size
+                            max_idsize = ceil((len(arr)+idptr_fh[i-1])*resize_by)
+                            max_embsize = ceil((emb.shape[0] + embptr_fh[i - 1]) *
+                                                                resize_by)
+                            resize = True
+                else:  # only check ID size conditions
+                    if len(arr) + idptr_fh[i - 1] > max_idsize or i >= len(idptr_fh):
+                        max_idsize = ceil((len(arr) + idptr_fh[i - 1] * resize_by))
                         resize = True
 
             # store the pointers that keep track of the start and
@@ -259,7 +265,7 @@ def _objects_to_embed(objs, fh, include_embedding_pointer=True):
                 id_fh = h5grp["id"]
                 if resize:
                     id_fh.resize((max_idsize,))
-                id_fh[idptr_fh[i - 1] : idptr_fh[i]] = arr
+                id_fh[idptr_fh[i - 1]: idptr_fh[i]] = arr
             else:
                 id_fh = h5grp.create_dataset(
                     "id", data=arr, maxshape=(None,), dtype=np.uint8, compression="gzip"
@@ -289,9 +295,9 @@ def _objects_to_embed(objs, fh, include_embedding_pointer=True):
                     embed_fh.resize(max_embsize, axis=0)
 
                 if include_embedding_pointer:
-                    embed_fh[embptr_fh[i - 1] : embptr_fh[i]] = emb
+                    embed_fh[embptr_fh[i - 1]: embptr_fh[i]] = emb
                 else:
-                    embed_fh[idptr_fh[i - 1] : idptr_fh[i]] = emb
+                    embed_fh[idptr_fh[i - 1]: idptr_fh[i]] = emb
 
             else:
                 embed_fh = h5grp.create_dataset(